+++ src/managementAPI.js
@@ -81,6 +81,7 @@ class ContentManagementClientImpl extends ContentDeliveryClientImpl {
       getTaxonomies: utils.bind(this.queryTaxonomies, this),
       queryTaxonomies: utils.bind(this.queryTaxonomies, this),
       queryTaxonomyCategories: utils.bind(this.queryTaxonomyCategories, this),
+      getTaxonomyCategory: utils.bind(this.getTaxonomyCategory, this),
     };
 
     logger.debug('ContentClient.create: Content Info:');